feat(setup): add default scripts to package manifest
