+++ src/setup/package-manifest.js
@@ -1,0 +1,48 @@
+const { readFile, writeFile } = require('fs-extra')
+
+const packagePropertySortOrder = [
+  'name',
+  'version',
+  'description',
+  'author',
+  'license',
+  'private',
+  'repository',
+  'keywords',
+  'main',
+  'types',
+  'scripts',
+  'peerDependencies',
+  'dependencies',
+  'devDependencies',
+]
+
+const addDefaultsToPackage = packageManifest => ({
+  ...packageManifest,
+  scripts: {
+    ...packageManifest.scripts,
+    build: 'mmm build',
+  },
+})
+
+const sortAndFilterProperties = packageManifest =>
+  Object.keys(packageManifest)
+    .filter(name => packagePropertySortOrder.includes(name))
+    .sort(
+      (a, b) =>
+        packagePropertySortOrder.indexOf(a) -
+        packagePropertySortOrder.indexOf(b),
+    )
+    .reduce((acc, name) => ({ ...acc, [name]: packageManifest[name] }), {})
+
+const updatePackageManifest = async packagePath => {
+  const packageManifest = addDefaultsToPackage(
+    JSON.parse(await readFile(packagePath, 'utf-8')),
+  )
+
+  const newPackage = sortAndFilterProperties(packageManifest)
+
+  await writeFile(packagePath, JSON.stringify(newPackage, null, 2) + '\n')
+}
+
+module.exports = updatePackageManifest
